requirements.txt
streamlit
PyMuPDF
opencv-python
pytesseract
Pillow
numpy
pandas
scikit-learn